+++ src/main/java/com/unipoint/merchant/dataaccess/MerchantUserDao.java
@@ -4,5 +4,5 @@ import com.unipoint.merchant.model.MerchantUser;
 
 public interface MerchantUserDao {
 	public MerchantUser validateLogin(String userName, String password);
-	public MerchantUser getMerchantUserById(int merchantUserRefId);
+	public MerchantUser getMerchantUserById(long merchantUserRefId);
 }
