+++ src/main/java/com/unipoint/merchant/dataaccess/OutletDao.java
@@ -6,7 +6,7 @@ import com.unipoint.merchant.model.Outlet;
 
 public interface OutletDao {   
 	public void addOutlet(Outlet outlet);
-	public Outlet getOutlet(int outletID);
-	public void deleteOutlet(int outletID);
+	public Outlet getOutlet(long outletID);
+	public void deleteOutlet(long outletID);
 	public List<Outlet> getAllOutlets();
 }
